feat: rename product surface to Concierge mode
- agent/concierge.py gate; one_room_control shim
- config orchestration.concierge_live_enabled (legacy frontdesk/one_room aliases)
- gateway/cli/tui import handle_concierge
- concierge_policy + control_plane concierge_mode_active (frontdesk_* aliases)
- tests: test_concierge + renamed policy suite

Product UX: enable with concierge_live_enabled true.
