+++ hermes_cli/config.py
@@ -987,10 +987,11 @@ DEFAULT_CONFIG = {
     # sessions (no live client) so accumulated agents don't pile up under memory
     # pressure. Reopening one re-resumes it from disk. 0/null disables.
     "max_live_sessions": 16,
-    # One-room control tower (opt-in): natural-language STOP/STATUS/WORKER gate
+    # Concierge mode (opt-in): natural-language STOP/STATUS/WORKER gate
     # that routes durable work through Kanban. Off by default.
+    # Legacy keys frontdesk_live_enabled / one_room_control_enabled still work.
     "orchestration": {
-        "frontdesk_live_enabled": False,
+        "concierge_live_enabled": False,
     },
     "agent": {
         "max_turns": 90,
