+++ tests/agent/test_control_plane.py
@@ -50,117 +50,117 @@ class TestIntentMapping:
 
     # WORKER_LANE policy, mode OFF → downgraded to NEW_TASK_MAIN / MAIN
     def test_worker_lane_policy_mode_off_downgrades_to_main(self):
-        d = classify("investigate the regression", frontdesk_mode_active=False)
+        d = classify("investigate the regression", concierge_mode_active=False)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.recommendation is Recommendation.MAIN
 
     def test_worker_lane_mode_off_downgrade_note_present(self):
-        d = classify("draft a report.md", frontdesk_mode_active=False)
+        d = classify("draft a report.md", concierge_mode_active=False)
         assert any("downgraded" in n for n in d.notes)
 
     def test_worker_lane_policy_mode_on_stays_worker(self):
-        d = classify("investigate the regression", frontdesk_mode_active=True)
+        d = classify("investigate the regression", concierge_mode_active=True)
         assert d.intent is Intent.NEW_TASK_WORKER
         assert d.recommendation is Recommendation.WORKER_LANE
 
     def test_worker_lane_artifact_mode_on(self):
-        d = classify("draft a report.md", frontdesk_mode_active=True)
+        d = classify("draft a report.md", concierge_mode_active=True)
         assert d.intent is Intent.NEW_TASK_WORKER
         assert d.recommendation is Recommendation.WORKER_LANE
 
     def test_worker_lane_code_edit_mode_on(self):
-        d = classify("refactor the module", frontdesk_mode_active=True)
+        d = classify("refactor the module", concierge_mode_active=True)
         assert d.intent is Intent.NEW_TASK_WORKER
         assert d.recommendation is Recommendation.WORKER_LANE
 
     # CONTROL + STOP signal
     def test_control_stop_mode_off(self):
-        d = classify("그만", frontdesk_mode_active=False)
+        d = classify("그만", concierge_mode_active=False)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_control_stop_mode_on(self):
-        d = classify("그만", frontdesk_mode_active=True)
+        d = classify("그만", concierge_mode_active=True)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_control_stop_en_mode_off(self):
-        d = classify("stop", frontdesk_mode_active=False)
+        d = classify("stop", concierge_mode_active=False)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_control_stop_en_mode_on(self):
-        d = classify("cancel", frontdesk_mode_active=True)
+        d = classify("cancel", concierge_mode_active=True)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     # CONTROL + ACK signal
     def test_control_ack_mode_off(self):
-        d = classify("고마워", frontdesk_mode_active=False)
+        d = classify("고마워", concierge_mode_active=False)
         assert d.intent is Intent.ACK
         assert d.recommendation is Recommendation.CONTROL
 
     def test_control_ack_mode_on(self):
-        d = classify("감사합니다", frontdesk_mode_active=True)
+        d = classify("감사합니다", concierge_mode_active=True)
         assert d.intent is Intent.ACK
         assert d.recommendation is Recommendation.CONTROL
 
     def test_control_ack_en_mode_off(self):
-        d = classify("thanks", frontdesk_mode_active=False)
+        d = classify("thanks", concierge_mode_active=False)
         assert d.intent is Intent.ACK
         assert d.recommendation is Recommendation.CONTROL
 
     def test_control_ack_en_mode_on(self):
-        d = classify("thank you", frontdesk_mode_active=True)
+        d = classify("thank you", concierge_mode_active=True)
         assert d.intent is Intent.ACK
         assert d.recommendation is Recommendation.CONTROL
 
     # MAIN + STATUS signal
     def test_main_status_signal(self):
-        d = classify("status?", frontdesk_mode_active=False)
+        d = classify("status?", concierge_mode_active=False)
         assert d.intent is Intent.STATUS
         assert d.recommendation is Recommendation.MAIN
 
     def test_main_status_signal_ko(self):
-        d = classify("지금 뭐 해?", frontdesk_mode_active=False)
+        d = classify("지금 뭐 해?", concierge_mode_active=False)
         assert d.intent is Intent.STATUS
         assert d.recommendation is Recommendation.MAIN
 
     # MAIN without STATUS → NEW_TASK_MAIN
     def test_main_no_status_is_new_task_main(self):
-        d = classify("이 파일 읽어줘", frontdesk_mode_active=False)
+        d = classify("이 파일 읽어줘", concierge_mode_active=False)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.recommendation is Recommendation.MAIN
 
     def test_main_default_conservative_en(self):
-        d = classify("explain the error message", frontdesk_mode_active=False)
+        d = classify("explain the error message", concierge_mode_active=False)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.recommendation is Recommendation.MAIN
 
     # STEER policy is mode-gated: OFF downgrades to MAIN; ON stays STEER.
     def test_steer_policy_mode_off_downgrades_to_main(self):
-        d = classify("근데 이거도 봐줘", frontdesk_mode_active=False)
+        d = classify("근데 이거도 봐줘", concierge_mode_active=False)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.recommendation is Recommendation.MAIN
 
     def test_steer_policy_mode_on_maps_to_steer_intent(self):
-        d = classify("근데 이거도 봐줘", frontdesk_mode_active=True)
+        d = classify("근데 이거도 봐줘", concierge_mode_active=True)
         assert d.intent is Intent.STEER
         assert d.recommendation is Recommendation.STEER
 
     def test_steer_policy_en_mode_on(self):
-        d = classify("also update the config file", frontdesk_mode_active=True)
+        d = classify("also update the config file", concierge_mode_active=True)
         assert d.intent is Intent.STEER
         assert d.recommendation is Recommendation.STEER
 
     # NOISE
     def test_noise_empty(self):
-        d = classify("", frontdesk_mode_active=False)
+        d = classify("", concierge_mode_active=False)
         assert d.intent is Intent.NOISE
         assert d.recommendation is Recommendation.CONTROL
 
     def test_noise_whitespace(self):
-        d = classify("   ", frontdesk_mode_active=False)
+        d = classify("   ", concierge_mode_active=False)
         assert d.intent is Intent.NOISE
         assert d.recommendation is Recommendation.CONTROL
 
@@ -170,63 +170,63 @@ class TestIntentMapping:
 # ---------------------------------------------------------------------------
 class TestTranscriptRenderVisibility:
     """STOP, NEW_TASK_WORKER, STEER, ACK, DUPLICATE, NOISE → transcript_render=True.
-    STATUS: True only when frontdesk_mode_active=True.
+    STATUS: True only when concierge_mode_active=True.
     NEW_TASK_MAIN → transcript_render=False.
     """
 
     def test_stop_transcript_render_true(self):
-        assert classify("그만", frontdesk_mode_active=False).transcript_render is True
+        assert classify("그만", concierge_mode_active=False).transcript_render is True
 
     def test_stop_transcript_render_true_mode_on(self):
-        assert classify("stop", frontdesk_mode_active=True).transcript_render is True
+        assert classify("stop", concierge_mode_active=True).transcript_render is True
 
     def test_new_task_worker_transcript_render_true(self):
-        d = classify("investigate the regression", frontdesk_mode_active=True)
+        d = classify("investigate the regression", concierge_mode_active=True)
         assert d.intent is Intent.NEW_TASK_WORKER
         assert d.transcript_render is True
 
     def test_steer_transcript_render_false_when_mode_off(self):
-        d = classify("근데 이거도 봐줘", frontdesk_mode_active=False)
+        d = classify("근데 이거도 봐줘", concierge_mode_active=False)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.transcript_render is False
 
     def test_steer_transcript_render_true_when_mode_on(self):
-        d = classify("근데 이거도 봐줘", frontdesk_mode_active=True)
+        d = classify("근데 이거도 봐줘", concierge_mode_active=True)
         assert d.intent is Intent.STEER
         assert d.transcript_render is True
 
     def test_steer_transcript_render_true_en_mode_on(self):
-        d = classify("also check the logs", frontdesk_mode_active=True)
+        d = classify("also check the logs", concierge_mode_active=True)
         assert d.intent is Intent.STEER
         assert d.transcript_render is True
 
     def test_ack_transcript_render_true(self):
-        d = classify("thanks", frontdesk_mode_active=False)
+        d = classify("thanks", concierge_mode_active=False)
         assert d.intent is Intent.ACK
         assert d.transcript_render is True
 
     def test_noise_transcript_render_true(self):
-        d = classify("", frontdesk_mode_active=False)
+        d = classify("", concierge_mode_active=False)
         assert d.intent is Intent.NOISE
         assert d.transcript_render is True
 
     def test_status_transcript_render_false_when_mode_off(self):
-        d = classify("status?", frontdesk_mode_active=False)
+        d = classify("status?", concierge_mode_active=False)
         assert d.intent is Intent.STATUS
         assert d.transcript_render is False
 
     def test_status_transcript_render_true_when_mode_on(self):
-        d = classify("status?", frontdesk_mode_active=True)
+        d = classify("status?", concierge_mode_active=True)
         assert d.intent is Intent.STATUS
         assert d.transcript_render is True
 
     def test_new_task_main_transcript_render_false(self):
-        d = classify("이 파일 읽어줘", frontdesk_mode_active=False)
+        d = classify("이 파일 읽어줘", concierge_mode_active=False)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.transcript_render is False
 
     def test_new_task_main_transcript_render_false_mode_on(self):
-        d = classify("이 파일 읽어줘", frontdesk_mode_active=True)
+        d = classify("이 파일 읽어줘", concierge_mode_active=True)
         assert d.intent is Intent.NEW_TASK_MAIN
         assert d.transcript_render is False
 
@@ -238,36 +238,36 @@ class TestStopSanctity:
     """classify('그만') always produces STOP / CONTROL regardless of mode."""
 
     def test_stop_invariant_ko_mode_off(self):
-        d = classify("그만", frontdesk_mode_active=False)
+        d = classify("그만", concierge_mode_active=False)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_stop_invariant_ko_mode_on(self):
-        d = classify("그만", frontdesk_mode_active=True)
+        d = classify("그만", concierge_mode_active=True)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_stop_invariant_en_mode_off(self):
-        d = classify("stop", frontdesk_mode_active=False)
+        d = classify("stop", concierge_mode_active=False)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_stop_invariant_en_mode_on(self):
-        d = classify("stop", frontdesk_mode_active=True)
+        d = classify("stop", concierge_mode_active=True)
         assert d.intent is Intent.STOP
         assert d.recommendation is Recommendation.CONTROL
 
     def test_stop_is_stop_property(self):
-        d = classify("그만", frontdesk_mode_active=False)
+        d = classify("그만", concierge_mode_active=False)
         assert d.is_stop is True
 
     def test_stop_is_control_property(self):
-        d = classify("그만", frontdesk_mode_active=False)
+        d = classify("그만", concierge_mode_active=False)
         assert d.is_control is True
 
     def test_stop_transcript_render_always_true(self):
         for mode in (False, True):
-            d = classify("그만", frontdesk_mode_active=mode)
+            d = classify("그만", concierge_mode_active=mode)
             assert d.transcript_render is True
 
 
@@ -278,27 +278,27 @@ class TestFingerprintDeterminism:
     """Same (text, mode) → same fingerprint; differs across mode bits."""
 
     def test_same_text_same_mode_same_fingerprint(self):
-        d1 = classify("investigate the regression", frontdesk_mode_active=False)
-        d2 = classify("investigate the regression", frontdesk_mode_active=False)
+        d1 = classify("investigate the regression", concierge_mode_active=False)
+        d2 = classify("investigate the regression", concierge_mode_active=False)
         assert d1.fingerprint == d2.fingerprint
 
     def test_different_mode_different_fingerprint(self):
-        d_off = classify("draft a report.md", frontdesk_mode_active=False)
-        d_on = classify("draft a report.md", frontdesk_mode_active=True)
+        d_off = classify("draft a report.md", concierge_mode_active=False)
+        d_on = classify("draft a report.md", concierge_mode_active=True)
         assert d_off.fingerprint != d_on.fingerprint
 
     def test_different_text_different_fingerprint(self):
-        d1 = classify("stop", frontdesk_mode_active=False)
-        d2 = classify("cancel", frontdesk_mode_active=False)
+        d1 = classify("stop", concierge_mode_active=False)
+        d2 = classify("cancel", concierge_mode_active=False)
         assert d1.fingerprint != d2.fingerprint
 
     def test_fingerprint_stored_on_decision(self):
-        d = classify("근데 이거도 봐줘", frontdesk_mode_active=False)
+        d = classify("근데 이거도 봐줘", concierge_mode_active=False)
         assert isinstance(d.fingerprint, str)
         assert len(d.fingerprint) > 0
 
     def test_fingerprint_in_to_dict(self):
-        d = classify("status?", frontdesk_mode_active=False)
+        d = classify("status?", concierge_mode_active=False)
         data = d.to_dict()
         assert "fingerprint" in data
         assert data["fingerprint"] == d.fingerprint
@@ -311,19 +311,19 @@ class TestHashability:
     """ControlPlaneDecision can be added to a set (frozen dataclass)."""
 
     def test_decision_hashable(self):
-        d = classify("그만", frontdesk_mode_active=False)
+        d = classify("그만", concierge_mode_active=False)
         s = {d}
         assert len(s) == 1
 
     def test_two_decisions_in_set(self):
-        d1 = classify("그만", frontdesk_mode_active=False)
-        d2 = classify("status?", frontdesk_mode_active=False)
+        d1 = classify("그만", concierge_mode_active=False)
+        d2 = classify("status?", concierge_mode_active=False)
         s = {d1, d2}
         assert len(s) == 2
 
     def test_equal_decisions_deduplicate_in_set(self):
-        d1 = classify("thanks", frontdesk_mode_active=False)
-        d2 = classify("thanks", frontdesk_mode_active=False)
+        d1 = classify("thanks", concierge_mode_active=False)
+        d2 = classify("thanks", concierge_mode_active=False)
         s = {d1, d2}
         assert len(s) == 1
 
@@ -349,11 +349,11 @@ class TestWorkerTaskSpecFromDecision:
     """Schema contracts for WorkerTaskSpec.from_decision (INV-7)."""
 
     def _worker_decision(self, text: str = "investigate and write report.md") -> ControlPlaneDecision:
-        return classify(text, frontdesk_mode_active=True)
+        return classify(text, concierge_mode_active=True)
 
     def test_title_defaults_to_first_60_chars(self):
         long_text = "investigate the regression in the authentication module and write report.md"
-        d = classify(long_text, frontdesk_mode_active=True)
+        d = classify(long_text, concierge_mode_active=True)
         spec = WorkerTaskSpec.from_decision(d)
         assert spec.title == long_text[:60].strip()
 
@@ -379,7 +379,7 @@ class TestWorkerTaskSpecFromDecision:
 
     def test_user_intent_is_raw_text(self):
         text = "investigate the regression and write report.md"
-        d = classify(text, frontdesk_mode_active=True)
+        d = classify(text, concierge_mode_active=True)
         spec = WorkerTaskSpec.from_decision(d)
         assert spec.user_intent == text
 
@@ -515,7 +515,7 @@ def _corpus_params_cp():
 def test_korean_corpus_row_matches_control_plane(row: dict):
     """Each corpus row's expected_intent and expected_recommendation match classify()."""
     mode = row["expected_intent"] in {"NEW_TASK_WORKER", "STEER"} or row["expected_recommendation"] in {"WORKER_LANE", "STEER"}
-    d = classify(row["text"], frontdesk_mode_active=mode)
+    d = classify(row["text"], concierge_mode_active=mode)
     assert d.intent.name == row["expected_intent"], (
         f"id={row['id']} text={row['text']!r} mode={mode} "
         f"expected={row['expected_intent']!r} got={d.intent.name!r}"
@@ -530,14 +530,14 @@ def test_korean_corpus_row_matches_control_plane(row: dict):
 # Mode-off downgrade is the ONLY pathway for non-MAIN/CONTROL under mode-off
 # ---------------------------------------------------------------------------
 class TestModeOffDowngradeInvariant:
-    """When frontdesk_mode_active=False, only MAIN and CONTROL recommendations appear."""
+    """When concierge_mode_active=False, only MAIN and CONTROL recommendations appear."""
 
     def test_corpus_mode_off_yields_only_main_or_control(self):
         """Mode-off downgrade: only MAIN / CONTROL recommendations may appear."""
         rows = _corpus()
         violations = []
         for row in rows:
-            d = classify(row["text"], frontdesk_mode_active=False)
+            d = classify(row["text"], concierge_mode_active=False)
             if d.recommendation not in {Recommendation.MAIN, Recommendation.CONTROL}:
                 violations.append(
                     f"id={row['id']} text={row['text']!r} got={d.recommendation.name}"
@@ -545,23 +545,39 @@ class TestModeOffDowngradeInvariant:
         assert violations == [], f"Mode-off yielded non-MAIN/CONTROL: {violations}"
 
     def test_explicit_worker_request_downgrades_when_mode_off(self):
-        d = classify("백그라운드로 돌려", frontdesk_mode_active=False)
+        d = classify("백그라운드로 돌려", concierge_mode_active=False)
         assert d.recommendation is Recommendation.MAIN
 
     def test_artifact_anchor_downgrades_when_mode_off(self):
-        d = classify("draft a report.md", frontdesk_mode_active=False)
+        d = classify("draft a report.md", concierge_mode_active=False)
         assert d.recommendation is Recommendation.MAIN
 
     def test_research_anchor_downgrades_when_mode_off(self):
-        d = classify("investigate the regression", frontdesk_mode_active=False)
+        d = classify("investigate the regression", concierge_mode_active=False)
         assert d.recommendation is Recommendation.MAIN
 
     def test_steer_downgrades_when_mode_off_and_stays_when_on(self):
-        d_off = classify("근데 이거도 봐줘", frontdesk_mode_active=False)
-        d_on = classify("근데 이거도 봐줘", frontdesk_mode_active=True)
+        d_off = classify("근데 이거도 봐줘", concierge_mode_active=False)
+        d_on = classify("근데 이거도 봐줘", concierge_mode_active=True)
         assert d_off.recommendation is Recommendation.MAIN
         assert d_off.intent is Intent.NEW_TASK_MAIN
         assert d_on.recommendation is Recommendation.STEER
 
         assert d_on.recommendation is Recommendation.STEER
         assert d_on.intent is Intent.STEER
+
+
+# ---------------------------------------------------------------------------
+# Temporary rename aliases
+# ---------------------------------------------------------------------------
+def test_legacy_frontdesk_mode_active_kw_alias():
+    """Legacy keyword still enables worker-lane routing."""
+    d_new = classify("draft a report.md", concierge_mode_active=True)
+    d_old = classify("draft a report.md", frontdesk_mode_active=True)
+    assert d_new.intent is d_old.intent
+    assert d_new.recommendation is d_old.recommendation
+    assert d_new.fingerprint == d_old.fingerprint
+    assert d_new.concierge_mode_active is True
+    assert d_old.frontdesk_mode_active is True
+    assert "concierge_mode_active" in d_new.to_dict()
+    assert "frontdesk_mode_active" in d_new.to_dict()
